refactored grpc common errors (#2200)
