+++ rpc/jsonrpc/server/http_server.go
@@ -17,7 +17,7 @@ import (
 	"golang.org/x/net/netutil"
 
 	"github.com/cometbft/cometbft/libs/log"
-	jsonRPCErrors "github.com/cometbft/cometbft/rpc/jsonrpc/errors"
+	grpcerr "github.com/cometbft/cometbft/rpc/grpc/errors"
 	types "github.com/cometbft/cometbft/rpc/jsonrpc/types"
 	cmttime "github.com/cometbft/cometbft/types/time"
 )
@@ -264,7 +264,7 @@ func (h maxBytesHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 func Listen(addr string, maxOpenConnections int) (listener net.Listener, err error) {
 	parts := strings.SplitN(addr, "://", 2)
 	if len(parts) != 2 {
-		return nil, jsonRPCErrors.ErrInvalidRemoteAddress{Addr: addr}
+		return nil, grpcerr.ErrInvalidRemoteAddress{Addr: addr}
 	}
 	proto, addr := parts[0], parts[1]
 	listener, err = net.Listen(proto, addr)
